backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]==3.1.7
python-multipart
python-dotenv
